+++ database/migrations/2021_03_15_102411_create_factures_table.php
@@ -20,7 +20,7 @@ class CreateFacturesTable extends Migration
             $table->boolean('payed')->default(false);
             $table->integer('price');
             $table->integer('payement_confirm_id')->nullable();
-            $table->integer('discord_msg_id')->nullable()->default(null);
+            $table->bigInteger('discord_msg_id')->nullable()->default(null);
             $table->string('service');
             $table->timestamps();
         });
